requirements.txt
Appium-Python-Client==1.1.0
appium-selector==1.3.2
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.4
djangorestframework==3.13.1
Faker==13.12.0
Flask==2.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
loging==0.0.1
MarkupSafe==2.1.1
numpy==1.22.4
pandas==1.4.2
Pillow==9.1.1
Pillow-PIL==0.1.dev0
pkg_resources==0.0.0
pure-python-adb==0.3.0.dev0
pyaes==1.6.1
pyasn1==0.4.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
random-user-agent==1.0.1
requests==2.27.1
rsa==4.8
selenium==3.141.0
selenium-stealth==1.0.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
telegram==0.0.1
Telethon==1.24.0
urllib3==1.26.9
webdriver-manager==3.7.0
Werkzeug==2.1.2
zipp==3.8.0
